+++ app/uploaders/asset_uploader.rb
@@ -1,0 +1,2 @@
+class AssetUploader < ApplicationUploader
+end
