+++ app/uploaders/image_uploader.rb
@@ -1,11 +1,2 @@
-class ImageUploader < CarrierWave::Uploader::Base
-  def store_dir
-    [store_root_dir, "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"].compact.join("/")
-  end
-
-  private
-
-  def store_root_dir
-    ENV["HEROKU_APP_NAME"] if ENV["USE_SHARED_S3_BUCKET"]
-  end
+class ImageUploader < ApplicationUploader
 end
