+++ db/migrate/20160421194616_add_custom_hue_css_url_to_sites.rb
@@ -1,0 +1,5 @@
+class AddCustomHueCssUrlToSites < ActiveRecord::Migration
+  def change
+    add_column :sites, :custom_hue_css, :string
+  end
+end
