+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160421081031) do
+ActiveRecord::Schema.define(version: 20160421194616) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -163,6 +163,7 @@ ActiveRecord::Schema.define(version: 20160421081031) do
     t.string   "category_title_format"
     t.boolean  "view_all",              default: false, null: false
     t.integer  "base_hue"
+    t.string   "custom_hue_css"
   end
 
   create_table "users", force: :cascade do |t|
